requirements.txt
Flask
numpy>=1.22,<1.24
opencv-python-headless
joblib
Pillow
tensorflow==2.12.0
gdown
scikit-learn